// src/types/FetchCartTypes.ts
export interface FetchCartItemType {
    id: number,
    title: string,
    description: string,
    price: number,
    discountPercentage: number;
    rating: number;
    stock: number;
    brand: string;
    category: string;
    thumbnail: string;
    images: string[],
}

export interface FetchInitialStateType {
    items: FetchCartItemType[];
    isLoading: boolean;
    error: string | null;
}